Backend/requirements.txt
# Core OCR + NLP
paddleocr==3.2.0
paddlepaddle>=2.6.1
spacy==3.8.7
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl

# PDF & Image processing
pdf2image==1.17.0
PyPDF2==3.0.1
camelot-py==0.11.0
opencv-python==4.11.0.86
Pillow==11.0.0
numpy>=1.26.0

# Fallback OCR
pytesseract==0.3.13

# Utilities
typing-extensions>=4.12.2
